Refatora fluxo de atendimento em blueprint
