+++ app/templates/atendimento/etapa10_outras_necessidades.html
@@ -67,7 +67,7 @@
         </button>
     </div>
     <div class="text-end">
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa10') }}">
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa10') }}">
             Próxima etapa
         </button>
     </div>
