+++ app/templates/atendimento/etapa11_atendimento.html
@@ -40,7 +40,7 @@
         <small id="avisoCesta" class="text-warning fst-italic {% if session['cadastro'].get('cesta_entregue') %}d-none{% endif %}">Nenhuma cesta básica será entregue neste atendimento.</small>
     </div>
     <div class="d-flex justify-content-between w-100">
-        <button type="submit" class="btn btn-success" id="btnFinalizar" data-next-url="{{ url_for('atendimento_etapa11') }}">Finalizar atendimento</button>
+        <button type="submit" class="btn btn-success" id="btnFinalizar" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa11') }}">Finalizar atendimento</button>
     </div>
 </form>
 {% endblock %}
