+++ app/templates/atendimento/etapa1_dados_pessoais.html
@@ -59,7 +59,7 @@
         </div>
     </div>
     <div class="text-end">
-        <button type="submit" class="btn btn-primary" id="btnProxima" disabled data-next-url="{{ url_for('atendimento_etapa1') }}">Próxima etapa</button>
+        <button type="submit" class="btn btn-primary" id="btnProxima" disabled data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa1') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
