+++ app/templates/atendimento/etapa2_endereco.html
@@ -66,8 +66,8 @@
 
     <!-- Botão -->
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa1') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa2') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa1') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa2') }}">Próxima etapa</button>
     </div>
 </form>
