+++ app/templates/atendimento/etapa4_contato.html
@@ -42,8 +42,8 @@
         <div id="email-feedback" class="invalid-feedback">Email inválido</div>
     </div>
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa3') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa4') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa3') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa4') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
