+++ app/templates/atendimento/etapa5_condicoes_habitacionais.html
@@ -97,8 +97,8 @@
         <input type="number" class="form-control" id="num_ventiladores" name="num_ventiladores" min="0" autocomplete="off" value="{{ session['cadastro'].get('num_ventiladores', '') }}">
     </div>
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa4') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa5') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa4') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa5') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
