+++ app/templates/atendimento/etapa6_saude_familiar.html
@@ -69,8 +69,8 @@
         </div>
     </div>
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa5') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa6') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa5') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa6') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
