+++ app/templates/atendimento/etapa8_renda_e_gastos.html
@@ -113,8 +113,8 @@
     </div>
     <input type="hidden" id="valor_aluguel_hidden" data-raw-value="{{ session['cadastro'].get('valor_aluguel', 0) }}">
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa7') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa8') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa7') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa8') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
