+++ app/templates/atendimento/etapa9_escolaridade.html
@@ -36,8 +36,8 @@
         <input type="text" class="form-control" id="curso_ou_serie_atual" name="curso_ou_serie_atual" autocomplete="off" value="{{ session['cadastro'].get('curso_ou_serie_atual', '') }}">
     </div>
     <div class="d-flex justify-content-between">
-        <a href="{{ url_for('atendimento_etapa8') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
-        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('atendimento_etapa9') }}">Próxima etapa</button>
+        <a href="{{ url_for('fluxo_atendimento.atendimento_etapa8') }}" class="btn btn-secondary" id="btnVoltar">Voltar</a>
+        <button type="submit" class="btn btn-primary" id="btnProxima" data-next-url="{{ url_for('fluxo_atendimento.atendimento_etapa9') }}">Próxima etapa</button>
     </div>
 </form>
 {% endblock %}
